\name{all.ie.themes}
\alias{all.ie.themes}
\title{List all themes}
\usage{
  all.ie.themes(x)
}
\value{
  Returns \code{NULL} but prints out list of Irish Census
  themes
}
\description{
  Basic list of Irish Census themes - use together with
  \code{theme.ie.tables} to identify tables of interest
}
\examples{
all.ie.themes()
}

