\name{theme.ie.tables}
\alias{theme.ie.tables}
\title{List all tables in a given theme}
\usage{
  theme.ie.tables(n)
}
\arguments{
  \item{n}{Theme number}
}
\value{
  Returns \code{NULL} but lists the tables in a given theme
}
\description{
  Shows all available tables in a given theme number
}
\examples{
theme.ie.tables(3)
}

